@@ -7,15 +7,13 @@
     "installcors": "gsutil.cmd cors set gcsbucket-cors-config.json  gs://historic_map_overlays",
     "builddocs": "pandoc -f markdown -t html --css=src/styles.css --self-contained --metadata pagetitle='User Guide' userdocs/userguide.md -o userdocs/userguide.html",
     "deploydocs": "gsutil.cmd cp userdocs/userguide.html gs://historic_map_overlays",
+    "installserver": "./buildserver.sh",
+    "server": "cd server && . Scripts/activate && python main.py",
     "ng": "ng",
-    "dev": "ng serve",
-    "build": "ng build --configuration=production",
+    "build": "cp -r src/assets/* server/app/assets/ && ng build --prod --configuration=production",
     "test": "ng test",
     "lint": "ng lint",
-    "predeploy": "npm run build",
-    "predeploy:dev": "npm run predeploy",
-    "deploy": "cd api && gcloud app deploy  --project yarrascrape --version production",
-    "deploy:dev": "cd api && app deploy --project google.com:intotecho --version yarraheritagemaps --no-promote",
+    "deploy": "cd server && . Scripts/activate && gcloud app deploy --quiet app.yaml",
     "compodoc": "npx compodoc -p src/tsconfig.app.json"
   },
   "private": true,
@@ -31,7 +29,7 @@
     "@angular/platform-browser": "^8.0.0",
     "@angular/platform-browser-dynamic": "^8.0.0",
     "@angular/router": "^8.0.0",
-    "@ngu/carousel": "^1.5.5",
+    "@openapitools/openapi-generator-cli": "^1.0.1-4.1.3",
     "@types/d3-scale": "^1.0.11",
     "@types/gapi": "^0.0.35",
     "@types/googlemaps": "^3.29.2",
@@ -73,7 +71,7 @@
     "rxjs-tslint": "^0.1.7",
     "ts-node": "~3.2.0",
     "tslint": "^5.17.0",
-    "typescript": "^3.4.5",
+    "typescript": "^3.5.3",
     "webpack": "^4.32.2"
   }
 }
