@@ -4,7 +4,6 @@ import { By } from '@angular/platform-browser';
 import { DebugElement } from '@angular/core';
 import { MatCardModule, MatExpansionModule } from '@angular/material';
 import { HeritageSiteInfoComponent } from './heritage-site-info.component';
-import { NguCarouselModule } from '@ngu/carousel';
 import { SoSService } from '../../../services/sos.service';
 import { HttpClientModule } from '@angular/common/http';
 import { HttpClient } from '@angular/common/http';
@@ -17,7 +16,7 @@ describe('HeritageSiteInfoComponent', () => {
     TestBed.configureTestingModule({
       declarations: [ HeritageSiteInfoComponent ],
       providers: [SoSService, HttpClient],
-      imports: [ MatExpansionModule, MatCardModule, NguCarouselModule, HttpClientModule]
+      imports: [ MatExpansionModule, MatCardModule, HttpClientModule]
     })
     .compileComponents();
   }));
