@@ -13,14 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 const BigQueryPublicAccessKey = 'AIzaSyArlp3qQv07_qjGr9P6fjHwy10HXeFj-Fw';
 const YarraScrapeOAuthClientID = '805117261004-oul8f6c5uhbvi4n043klqpj5u7jd74co.apps.googleusercontent.com';
 
 export const environment = {
-  production: true,
-  authClientID: YarraScrapeOAuthClientID,
+  production: false,
+  authClientID: '419125973937-kl2cru5pu2vfugne7lr1hosgseh4lo1s.apps.googleusercontent.com',
   authScope: 'https://www.googleapis.com/auth/bigquery',
   projectId: 'yarrascrape',
-  dataAPIBasePath: 'https://yarrascrape.appspot.com',
-  _DATASET: 'YarraPlanning' // not used.
+  dataAPIBasePath: 'http://localhost:8080',
+  _DATASET: 'YarraPlanning',
+  BigQueryPublicAccessKey: 'AIzaSyArlp3qQv07_qjGr9P6fjHwy10HXeFj-Fw'
 };
